refactor: Issue 도메인에서 Comment 도메인 분리
