+++ BE/src/main/java/com/team31/codesquad/issuetracker/service/IssueService.java
@@ -1,7 +1,6 @@
 package com.team31.codesquad.issuetracker.service;
 
 import com.team31.codesquad.issuetracker.dto.OpenClosedCountResult;
-import com.team31.codesquad.issuetracker.dto.comment.CommentCreateRequest;
 import com.team31.codesquad.issuetracker.dto.issue.IssueAssigneesChangeRequest;
 import com.team31.codesquad.issuetracker.dto.issue.IssueCreateRequest;
 import com.team31.codesquad.issuetracker.dto.issue.IssueCreateResponse;
@@ -11,7 +10,6 @@ import com.team31.codesquad.issuetracker.dto.issue.IssueMilestoneChangeRequest;
 import com.team31.codesquad.issuetracker.dto.issue.IssueResponse;
 import com.team31.codesquad.issuetracker.dto.issue.IssueStatusChangeRequest;
 import com.team31.codesquad.issuetracker.dto.issue.MultiIssueStatusChangeRequest;
-import com.team31.codesquad.issuetracker.dto.reaction.ReactionCreateRequest;
 import java.util.List;
 
 public interface IssueService {
@@ -24,8 +22,6 @@ public interface IssueService {
 
     IssueDetailResponse getIssue(Long issueId);
 
-    Long createComment(Long issueId, CommentCreateRequest request);
-
     void changeStatus(Long issueId, IssueStatusChangeRequest request);
 
     void changIssuesStatus(MultiIssueStatusChangeRequest request);
@@ -36,8 +32,4 @@ public interface IssueService {
 
     void changeLabels(Long issueId, IssueLabelsChangeRequest request);
 
-    Long createReaction(Long issueId, Long commentId, String loginName,
-            ReactionCreateRequest request);
-
-    void deleteReaction(Long reactionId);
 }
