+++ BE/src/main/java/com/team31/codesquad/issuetracker/web/controller/IssueController.java
@@ -2,7 +2,6 @@ package com.team31.codesquad.issuetracker.web.controller;
 
 import com.team31.codesquad.issuetracker.config.mvc.annotation.LoginName;
 import com.team31.codesquad.issuetracker.dto.OpenClosedCountResult;
-import com.team31.codesquad.issuetracker.dto.comment.CommentCreateRequest;
 import com.team31.codesquad.issuetracker.dto.issue.IssueAssigneesChangeRequest;
 import com.team31.codesquad.issuetracker.dto.issue.IssueCreateRequest;
 import com.team31.codesquad.issuetracker.dto.issue.IssueCreateResponse;
@@ -12,7 +11,6 @@ import com.team31.codesquad.issuetracker.dto.issue.IssueMilestoneChangeRequest;
 import com.team31.codesquad.issuetracker.dto.issue.IssueResponse;
 import com.team31.codesquad.issuetracker.dto.issue.IssueStatusChangeRequest;
 import com.team31.codesquad.issuetracker.dto.issue.MultiIssueStatusChangeRequest;
-import com.team31.codesquad.issuetracker.dto.reaction.ReactionCreateRequest;
 import com.team31.codesquad.issuetracker.service.IssueService;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -61,13 +59,6 @@ public class IssueController {
         return issueService.getIssue(issueId);
     }
 
-    @PostMapping("/api/v1/issues/{issueId}/comments")
-    public ResponseEntity<Long> createComment(@PathVariable Long issueId,
-            @Validated @RequestBody CommentCreateRequest request) {
-        Long createdCommentId = issueService.createComment(issueId, request);
-        return ResponseEntity.status(HttpStatus.CREATED).body(createdCommentId);
-    }
-
     @PatchMapping("/api/v1/issues/{issueId}/status")
     public ResponseEntity<Void> changeStatus(@PathVariable Long issueId,
             @Validated @RequestBody IssueStatusChangeRequest request) {
@@ -103,18 +94,4 @@ public class IssueController {
         return ResponseEntity.noContent().build();
     }
 
-    @PostMapping("/api/v1/issues/{issueId}/comments/{commentId}/reactions")
-    public ResponseEntity<Long> createReaction(
-            @PathVariable Long issueId, @PathVariable Long commentId,
-            @LoginName String loginName, @RequestBody ReactionCreateRequest request) {
-        Long createdReactionId = issueService.createReaction(issueId, commentId, loginName,
-                request);
-        return ResponseEntity.status(HttpStatus.CREATED).body(createdReactionId);
-    }
-
-    @DeleteMapping("/api/v1/issues/comments/reactions/{reactionId}")
-    public ResponseEntity<Void> deleteReaction(@PathVariable Long reactionId) {
-        issueService.deleteReaction(reactionId);
-        return ResponseEntity.noContent().build();
-    }
 }
